Remove JSON Data plugin for dataframes
Renamed dataframe.hdf5 to dataframe.frame

The reason for this is that the JSON format has some edge cases
that it can't serialize. In addition data types are limited to what
can be serialized to the JSON attributes in the Aiida database

The performance of the HDF5 plugin is a slower by a factor 3 for small
data frames (4 rows x 5 columns) but faster by orders of magnitude for
large dataframes (10000 rows x 52 columns)

Bonus the plugin can now be compatible with AiiDA 1.X :)
