+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ jobs:
     timeout-minutes: 30
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10"]
+        python-version: ["3.7","3.8", "3.9", "3.10"] #3.7 will install AiiDA 1.X
         aiida-version: ["stable"]
 
     services:
