+++ tests/test_data.py
@@ -13,7 +13,6 @@ from aiida.plugins import DataFactory
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_roundtrip(entry_point):
@@ -48,7 +47,6 @@ def test_roundtrip(entry_point):
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_multiindex_columns_roundtrip(entry_point):
@@ -82,7 +80,6 @@ def test_multiindex_columns_roundtrip(entry_point):
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_multiindex_index_roundtrip(entry_point):
@@ -107,7 +104,6 @@ def test_multiindex_index_roundtrip(entry_point):
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_query_columns(entry_point):
@@ -158,7 +154,6 @@ def test_query_columns(entry_point):
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_query_multiindex_columns(entry_point):
@@ -211,7 +206,6 @@ def test_query_multiindex_columns(entry_point):
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_query_index(entry_point):
@@ -255,33 +249,6 @@ def test_query_index(entry_point):
 
 
 @pytest.mark.parametrize("entry_point", ("dataframe.frame",))
-def test_non_serializable(entry_point):
-    """
-    Make sure that non-serializable things are caught
-    early before storing the Dataframe
-    """
-
-    # Example from pandas Docs
-    df = pd.DataFrame(
-        {
-            "A": 1.0,
-            "B": pd.Timestamp("20130102"),
-            "C": pd.Series(1, index=list(range(4)), dtype="float32"),
-            "D": np.array([3] * 4, dtype="int32"),
-            "E": pd.Categorical(["test", "train", "test", "train"]),
-            "F": "foo",
-            "G": 1 + 2j,
-        }
-    )
-
-    PandasFrameData = DataFactory(entry_point)
-
-    # Complex numbers can not be serialized/deserialized in a consistent manner
-    with pytest.raises(TypeError):
-        PandasFrameData(df)
-
-
-@pytest.mark.parametrize("entry_point", ("dataframe.hdf5",))
 def test_complex_hdf5(entry_point):
     """
     Make sure that non-serializable things are caught
@@ -311,7 +278,7 @@ def test_complex_hdf5(entry_point):
     assert_frame_equal(loaded.df, df)
 
 
-@pytest.mark.parametrize("entry_point", ("dataframe.hdf5",))
+@pytest.mark.parametrize("entry_point", ("dataframe.frame",))
 def test_nan_values_hdf5(entry_point):
     """
     Make sure that NaNs are eliminated (transformed to None)
@@ -339,48 +306,10 @@ def test_nan_values_hdf5(entry_point):
     assert_frame_equal(loaded.df, df)
 
 
-@pytest.mark.parametrize("entry_point", ("dataframe.frame",))
-def test_nan_values(entry_point):
-    """
-    Make sure that NaNs are eliminated (transformed to None)
-    before being stored in the Database
-    """
-
-    df = pd.DataFrame(
-        {
-            "None": [
-                np.NAN,
-                float("NaN"),
-                np.inf,
-                float("inf"),
-            ],
-        }
-    )
-
-    # The NaNs are reconstructed only in the final Dataframe
-    # by pandas noticing the None values in a numeric column
-    df_after = pd.DataFrame(
-        {
-            "None": [float("NaN"), float("NaN"), float("NaN"), float("NaN")],
-        }
-    )
-
-    PandasFrameData = DataFactory(entry_point)
-    node = PandasFrameData(df)
-
-    # Make sure that NaNs and infs already disappeared during this step
-    assert all(row["None"] is None for row in node.obj.as_dict()["data"])
-    node.store()
-
-    loaded = load_node(node.pk)
-    assert_frame_equal(loaded.df, df_after)
-
-
 @pytest.mark.parametrize(
     "entry_point",
     (
         "dataframe.frame",
-        "dataframe.hdf5",
     ),
 )
 def test_wrong_inputs(entry_point):
